third_party/blink/web_tests/fast/events/tabindex-no-focusable-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS document.activeElement.id is "MoveToMe"
PASS document.activeElement.id is "MoveToMe"
   
